Done
The core idea, or at least as good as it gets with my skill atm, is to provide interface that dynamically knows namespace name and key names from the JSON file and adds them accordingly to the view. 

The only problem, is that I haven't fond a reliable way to assign key value from the view to the controller so it is independant from the view. 

As it stands right now, when you modify position of the JSON key in the array and/or add new JSON key, you will have to manually put the key value inside the view.